% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\title{Add default columns to a data.frame}
\usage{
defaults(x, def)
}
\arguments{
\item{x}{Data.frame that serves as input}

\item{def}{Data.frame that contains the defaults}
}
\value{
Data.frame that is filled with the values in \code{x} and/or \code{def}
}
\description{
Add default columns to a data.frame
}
