% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{discretize}
\alias{discretize}
\title{Discretize space}
\usage{
discretize(x, min_x = min(x), dx = 0.1)
}
\arguments{
\item{x}{Numeric vector containing values to be discretized.}

\item{min_x}{Value to add to \code{x} so that it starts at 0. Needed to ensure
that those values of \code{x} that are minimal are assigned to cell 0. 
Defaults to the minimum of \code{x}.}

\item{dx}{Numeric denoting the size of a cell. Defaults to \code{0.1}.}
}
\value{
Integer vector containing indices of the cells in which \code{x} can 
be attributed.
}
\description{
Takes in a vector of continuous data and then assigns the observed positions
to a cell in a grid based on the discretization \code{dx} parameter. This 
function is needed to translate the continuous positions that \code{predped} 
provides to discrete cells that \code{QVEmod} uses.
}
\details{
Output consists of the indices of the cells that the continuous position lies 
in. Note that indexing starts at 0 instead of 1 (following Python indexing).
}
