% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hill_function}
\alias{hill_function}
\title{Hill function}
\usage{
hill_function(exposure, alpha = 0.332, lambda = 10^(6.8))
}
\arguments{
\item{exposure}{Numeric denoting the total exposure an agent has had to an 
infectious agent, or otherwise the ligand concentration.}

\item{alpha}{Numeric denoting the Hill coefficient. Defaults to \code{0.332},
which was taken from code originally written by the QVEmod developers.}

\item{lambda}{Numeric denoting the ligand concentration producing half 
occupation. Defaults to \code{10^(6.8)}, which was taken from code originally
written by the QVEmod developers.}
}
\value{
Numeric denoting the risk of infection
}
\description{
Hill function as defined by Busra Atamer Balkan and Colin Teberg. Used as 
one of the potential dose-response functions when computing infection risk.
}
