% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_risk.R
\name{infection_risk}
\alias{infection_risk}
\title{Compute an agent's infection risk}
\usage{
infection_risk(
  data,
  time_step = 0.5,
  average_emission = 10^6,
  surface_exposure_ratio = 0.01,
  fx = hill_function,
  ...
)
}
\arguments{
\item{data}{Results coming from the \code{\link[beprepared]{simulate}} 
function. Should in the least contain the agent characteristics under 
\code{"agents"} and the exposure of the agents under \code{"agent_exposure"}.}

\item{time_step}{Numeric denoting the time between each iteration. Defaults 
to \code{0.5} (the same as in \code{\link[predped]{simulate,predped-method}}).}

\item{surface_exposure_ratio}{}

\item{fx}{Function that will translate contamination exposure to risk. 
By default uses the \code{\link[predped]{hill_function}}.}

\item{...}{Additional arguments passed on to \code{fx}.}
}
\value{
Data.frame containing the agent and their exposure/risk
}
\description{
Compute an agent's infection risk
}
