% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{relative_movement}
\alias{relative_movement}
\title{Compute movement relative to previous position}
\usage{
relative_movement(x)
}
\arguments{
\item{x}{Numeric vector containing values that should be made relative to 
each other.}
}
\value{
Numeric vector containing movements relative to the previous ones
}
\description{
Compute movement relative to previous position
}
