% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{segmentize}
\alias{segmentize}
\title{Transform object to segments}
\usage{
segmentize(object, discretize = TRUE, origin = c(0, 0), dx = 0.1)
}
\arguments{
\item{object}{Instance of the \code{\link[predped]{object-class}} to be 
segmentized}

\item{discretize}{Logical denoting whether the positions should be discretized.
If \code{TRUE}, segments will refer to indices of cells instead of absolute 
continuous positions in space. Defaults to \code{TRUE}.}

\item{origin}{X- and y-coordinate that represents the bottom-left corner of 
the space in which the object is located. Allows for accurate discretization.
Ignored if \code{discretize = FALSE}. Defaults to \code{c(0, 0)}.}

\item{dx}{Numeric denoting the size of a cell. Ignored if 
\code{discretize = FALSE}. Defaults to \code{0.1}.}
}
\value{
Data.frame containing the beginning (\code{"x1"} and \code{"y1"}) and
end coordinates (\code{"x2"} and \code{"y2"}) of the object.
}
\description{
Takes in an instance of the \code{\link[predped]{object-class}} and returns 
a data.frame of segment points for this object. Used for translation purposes
from \code{predped} to \code{QVEmod}.
}
