% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap}
\alias{heatmap}
\title{Create a heatmap}
\usage{
heatmap(
  data,
  heatmap.fill = c("white", "salmon"),
  Z.label = "Z",
  Z.limits = c(0, max(data$Z)),
  X.limits = range(data$X),
  Y.limits = range(data$Y),
  shape.color = "black",
  shape.fill = "grey",
  shape.linewidth = 1,
  plot.title = " ",
  plot.title.size = 10,
  plot.title.hjust = 0.5,
  axis.title.size = 10,
  axis.text.size = 8,
  legend.position = "bottom",
  legend.title.size = axis.title.size,
  legend.text.size = axis.text.size,
  dark_mode = FALSE,
  ...
)
}
\arguments{
\item{data}{Data.frame containing values for \code{X}, \code{Y}, and \code{Z}
where \code{X} and \code{Y} denote the coordinates in 2D space and \code{Z}
denotes the value assigned to this cell.}

\item{heatmap.fill}{Character vector containing two colors to use as 
extremes for the heatmap. If \code{dark_mode = TRUE}, the first of these 
colors is changed to \code{"black"}. Defaults to 
\code{c("white", "cornflowerblue")}.}

\item{Z.label}{Character denoting the label of the variable \code{Z} to be 
used in the legend. Defaults to \code{"Z"}.}

\item{Z.limits}{Numeric vector containing the limits of the heatmap in the 
Z dimension. Defaults to the minimum and maximum value of \code{Z} in the 
\code{data}.}

\item{X.limits, Y.limits}{Numeric vector containing the limits of the heatmap
in the X and Y dimensions, serving as limits to the plotted figure. Defaults 
to the minimum and maximum value of \code{X} and \code{Y} in the \code{data}.}

\item{shape.color}{Character denoting the color of the boundary of the 
setting in which agents are walking around. Defaults to \code{"black"} when 
\code{dark_mode = FALSE}, and to \code{"white"} otherwise.}

\item{shape.fill}{Character denoting the fill color of the setting in which 
the agents are running around. Defaults to \code{"grey"} when 
\code{dark_mode = FALSE} and to \code{"black"} otherwise.}

\item{shape.linewidth}{Numeric denoting the width of the boundary of the 
setting in which agents are walking around. Defaults to \code{1}.}

\item{plot.title}{Character denoting the title fo the plot. Defaults to an 
empty string.}

\item{plot.title.size}{Numeric denoting the text size of the plot title.
Defaults to \code{10}.}

\item{plot.title.hjust}{Numeric denoting the position of the plot title, with
\code{0} coding for left, \code{1} for right, and \code{0.5} for the middle.
Defaults to \code{0.5}.}

\item{axis.title.size}{Numeric of the text size of the axis title. Defaults
to \code{10}.}

\item{axis.text.size}{Numeric denoting the text size of the axis text.
Defaults to \code{8}.}

\item{legend.position}{Character denoting where the legend should be located.
Defaults to \code{"right"}.}

\item{legend.title.size}{Numeric denoting the size of the legend title.
Defaults to the size of \code{axis.title.size}.}

\item{legend.text.size}{Numeric denoting the size of the legend text. 
Defaults to the size of \code{axis.text.size}.}

\item{dark_mode}{Logical that can toggle the default colorpallette of predped's
dark mode. Defaults to \code{FALSE}.}

\item{...}{Additional arguments that remain unused, but allow for the use 
of the same set of arguments across plotting functions.}
}
\value{
Heatmap
}
\description{
Create a heatmap
}
\details{
Returns a heatmap created through \code{ggplot2}. Used to display the spread
of contamination through space. For simplicity sake, makes use of the same 
argument as the \code{\link[predped]{plot-method}} of \code{predped}, allowing
us to simply use one set of arguments for all plots simultaneously.
}
