% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.R
\name{hill_function}
\alias{hill_function}
\title{Hill function}
\usage{
hill_function(exposure, alpha, lambda)
}
\arguments{
\item{exposure}{Numeric denoting the total exposure an agent has had to an 
infectious agent, or otherwise the ligand concentration.}

\item{alpha}{Numeric denoting the Hill coefficient.}

\item{lambda}{Numeric denoting the ligand concentration producing half 
occupation.}
}
\value{
Numeric denoting the risk of infection
}
\description{
Hill function as defined by Busra Atamer Balkan and Colin Teberg. Used as 
one of the potential dose-response functions when computing infection risk.
}
