% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate agent movements}
\usage{
simulate(
  filename,
  environment,
  archetypes = "BaselineEuropean",
  weights = rep(1/length(archetypes), length(archetypes)),
  dx = 0.1,
  save_gif = FALSE,
  plot_args = list(),
  env_args = data.frame(decay_rate_air = 1.51, decay_rate_droplet = 0.3,
    decay_rate_surface = 0.262, air_exchange_rate = 0.2, droplet_to_surface_transfer_rate
    = 18.18),
  surf_args = data.frame(prob = 1, transfer_decay_rate = 0.5, touch_frequency = 15,
    surface_decay_rate = 0.969),
  item_args = data.frame(prob = 1, transfer_decay_rate = 0.7, surface_ratio = 0.5,
    surface_decay_rate = 0.274),
  agent_args = data.frame(prob = rep(1/3, 3), viral_load = c(1, 0, 0),
    contamination_load_air = c(0, 0, 0), contamination_load_droplet = c(0, 0, 0),
    contamination_load_surface = c(1, 0, 0), emission_rate_air = rep(0.53, 3),
    emission_rate_droplet = rep(0.47, 3), pick_up_air = c(2.3, 30, 30), pick_up_droplet =
    c(2.3, 30, 30), wearing_mask = c(0, 0, 0)),
  env_config = data.frame(AirCellSize = 50, MobilityCellSize = 100 * dx, AgentReach = 50,
    SimulationTimeStep = time_step, HandwashingContaminationFraction = 0.3,
    HandwashingEffectDuration = 0.5, MaskEmissionAerosolReductionEfficiency = 0.4,
    MaskEmissionDropletReductionEfficiency = 0.04, MaskAerosolProtectionEfficiency = 0.4,
    MaskDropletProtectionEfficiency = 0.04, CleaningInterval = 1, Diffusivity = 23,
    WallAbsorbingProportion = 0, CoughingRate = 121, CoughingFactor = 1,
    CoughingAerosolPercentage = 1, CoughingDropletPercentage = 1, 
    
    SurfaceExposureRatio = 0.01),
  output_config = data.frame(Suppress = FALSE, Path = file.path("results", "output"),
    AerosolContaminationWriteInterval = 15, AerosolContaminationPrecision = 17,
    DropletContaminationWriteInterval = 15, DropletContaminationPrecision = 17,
    SurfaceContaminationWriteInterval = 15, SurfaceContaminationPrecision = 17),
  time_step = 0.5,
  ...
)
}
\arguments{
\item{filename}{Character denoting what to call this simulation in the output
files.}

\item{environment}{Object of the \code{\link[predped]{background-class}} 
denoting the environment in which agents are to walk around.}

\item{archetypes}{Character vector denoting the parameter sets of 
\code{predped} to use for the simulation. Defaults to "BaselineEuropean".}

\item{weights}{Numeric vector denoting the probability with which each 
parameter set can be chosen to enter the simulation. Defaults to equal 
weighing of each entry in \code{archetypes}.}

\item{dx}{Numeric denoting the width and height of each cell in the simulation.
While \code{predped} operates on continuous space, \code{QVEmod} requires 
this space to consist of discretized cells. Defaults to \code{0.1}, which 
should be interpreted as 10cm (i.e., the units of the \code{environment} are 
in meters). This default is based on the contagion patterns of \code{QVEmod}
which have been parametrized on 10cm. Note that changing this argument is 
possible and will influence the \code{MobilityCellSize} argument in the 
\code{env_config} defaults.}

\item{save_gif}{Logical denoting whether to save a gif of the simulation. 
Only shows the movement of the agents, not whether they are spreading a 
disease. Defaults to \code{FALSE}.}

\item{plot_args}{Named list containing optional arguments to use when 
plotting the simulation. Only used when \code{save_gif = TRUE}. Defaults to 
an empty list.}

\item{env_args}{Data.frame containing optional arguments for the parameters 
that make up the \code{Environment} as defined by QVEmod. Defaults to the 
values used by default through QVEmod.}

\item{surf_args}{Data.frame containing optional arguments for the parameters 
that make up the \code{Fixture}s in the environment, as defined by QVEmod. 
Defaults to the values used by default through QVEmod.}

\item{item_args}{Data.frame containing optional arguments for the parameters 
that make up the \code{Item}s in the environment, as defined by QVEmod. 
Defaults to the values used by default through QVEmod.}

\item{agent_args}{Data.frame containing optional arguments for the parameters 
that make up the \code{Agent}s in the simulation, as defined by QVEmod. 
Defaults to the values used by default through QVEmod.}

\item{env_config}{Data.frame containing optional arguments for the 
configuration of the environment. Defaults to the values found in the 
`default_config.json` file of QVEmod.}

\item{output_config}{Data.frame containing optional arguments for the 
configuration of the output Defaults to the values found in the 
`default_config.json` file of QVEmod.}

\item{...}{Additional arguments passed on to 
\code{\link[predped]{simulate,predped-method}}}
}
\value{
Return \code{NULL}. Will save all of the output in different files
according to the filename provided in the function.
}
\description{
Use the \code{\link[predped]{simulate,predped-method}} function to simulate 
agent movements and assess the risk of infection for each of the agents in 
the simulation. Combines the different steps that you would normally have 
to go through in one function, allowing one to simulate data with minimal 
effort.
}
